third_party/blink/web_tests/accessibility/table-with-th-role-gridcell-crash.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<style>
tr {
    display: grid;
  }
</style>
<table role="grid">
  <th id="cell1" role="gridcell"></th>
</table>

<script>
function axElementById(id) {
    return accessibilityController.accessibleElementById(id);
}

test(function(t) {
    var ax = axElementById("cell1");
    assert_equals(ax.role, "AXRole: AXGridCell");
}, "A <th role='gridcell'> in a table with extra layout objects doesn't crash");
</script>
<!--

<script>
function axElementById(id) {
    return accessibilityController.accessibleElementById(id);
}
</script>
-->
